src/main/java/dev/tobee/telegram/request/message/EditMessageMedia.java
package dev.tobee.telegram.request.message;

import java.util.Map;
import java.util.Optional;

import com.fasterxml.jackson.core.type.TypeReference;
import dev.tobee.telegram.model.message.ResponseWrapper;
import dev.tobee.telegram.request.Request;
import dev.tobee.telegram.request.body.EditMessageMediaBody;
import dev.tobee.telegram.util.DefaultJsonMapper;
import dev.tobee.telegram.util.DefaultObjectMapper;

public class EditMessageMedia implements Request<ResponseWrapper<Boolean>> {
    private static final String METHOD = "editMessageMedia";
    private static final TypeReference<ResponseWrapper<Boolean>> reference = new TypeReference<>() {};
    private final EditMessageMediaBody body;

    public EditMessageMedia(EditMessageMediaBody body) {
        this.body = body;
    }

    @Override
    public String getMethod() {
        return METHOD;
    }

    @Override
    public TypeReference<ResponseWrapper<Boolean>> getResponseType() {
        return reference;
    }

    @Override
    public Optional<Map<Object, Object>> getBody() {
        Map<Object, Object> bodyMap = DefaultObjectMapper.convertToMap(body);
        DefaultJsonMapper.convertMapValueToStringJson(bodyMap, "reply_markup", body.replyMarkup());
        return Optional.of(bodyMap);
    }
}
